pyproject.toml
[project]
name = "tg-presentation-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=22.6",
]

[dependency-groups]
dev = [
    "ruff>=0.15.1",
]
